Refactor XML model components: consolidate ManifestHeader and ManifestRef definitions, remove unused imports
